requirements.txt
asgiref==3.6.0
autopep8==2.0.2
black==23.1.0
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-import-export==3.1.0
et-xmlfile==1.1.0
MarkupPy==1.14
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pycodestyle==2.10.0
PyYAML==6.0
python-dotenv==1.0.0
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
xlrd==2.0.1
xlwt==1.3.0